Updated printing grids data structure info

+++ include/mesh.h
@@ -47,7 +47,7 @@ typedef struct {
 	Topo	*topo; //Topology info
 	int	id; // grid id; 0:finest, ngrids-1:coarsest
 	int	n[MAX_DIMENSION]; // Grids in each direction
-	double	range[MAX_DIMENSION][2]; // Range of grid points
+	int	range[MAX_DIMENSION][2]; // Range of grid points
 	double	**coord; // Coordinates in each direction
 	double	h; // Grid characteristic length
 	double	para[4]; // Domain splitting quality measures
